debian: use new Google keys for loopback server
These keys have been generated in the Google API Console for a "desktop
application". This also means that we cannot use an external OAuth
callback, but we need to use a loopback URL. Luckily, Google does not
really check whether there's actually a server responding on this
address, so we can get away without implementing one.

This seems to remove the warning on the insecure browser.

Fixes: https://github.com/ubports/account-plugins/issues/40